feat: :fire: Modulo Style
💄 feat: Add header and icons

Added the header and icons to the wallet.

🎨 style: Improve header layout in CSS

Improved the header layout using CSS in the wallet.

🎨 style: Add all icons and CSS styling for index page

Added all icons and applied CSS styling for the first page (index) of the wallet.

🎨 style: Add CSS with colors for text and footer

Added CSS styles with colors extracted from the text and applied them to the footer section of the wallet.
This enhancement improves the visual appeal and consistency of the wallet's design, ensuring a cohesive user experience across the application.

✨ feat: Add fonts and footer

Added fonts to enhance typography and implemented the footer section in the wallet.

🔖 feat: Adding footer

Implemented the footer section in the wallet.

🐛 fix: Fix small error

Fix small error in CSS

🐛 fix: Fix footer and scroll divs

Fixed issues with the footer display and scroll behavior in the wallet.

🐛 fix: Fix some errors in details.css and style.css

Addressed several errors in the details.css and style.css files of the wallet interface.

🎨 style: Update Figma style

Modified the style of the wallet interface to match the design in Figma.

🎨 style: Update color scheme

Changed the color scheme used in the wallet interface.

🎉 feat: Finalize wallet design

Completed the design adjustments in style.css, balance.css, and details.css to finalize the wallet interface.