release: :bookmark: Modulo Balance
✨ feat: Add page details and balance section

Implemented functionality to display page details and balance information in the wallet.

✨ feat: Add balance page

Implemented the addition of the balance page to the wallet interface.

✨ feat: Add balance parameters

Implemented the addition of balance parameters to the wallet interface.

🎨 style: Update CSS for the balance page

Made necessary changes to the CSS styles for the balance page of the wallet interface.

🎨 style: Revise style for balance page

Updated the CSS styling for the balance page to enhance its appearance and layout.

✨ feat: Finish balance page

Completed the implementation of the balance page, ensuring all features and functionalities are in place.

🚀 feat: Finalize balance page implementation

Completed the implementation of the balance page and addressed any remaining issues.

🐛 fix: Fix some errors in balance.css

Addressed several errors in the balance.css file of the wallet interface.

🎨 style: Update Figma style

Modified the style of the wallet interface to match the design in Figma.

🎨 style: Update color scheme

Changed the color scheme used in the wallet interface.

🎨 style: Update fonts

Changed the fonts used in the wallet interface.

🎨 style: Update color palette

Adjusted the colors used throughout the wallet interface.

🎉 feat: Finalize wallet design

Completed the design adjustments in style.css, balance.css, and details.css to finalize the wallet interface.

:fire: feat: Modulo balance html

✨ feat: Add page details and balance section

Implemented functionality to display page details and balance information in the wallet.

✨ feat: Add balance page

Implemented the addition of the balance page to the wallet interface.

✨ feat: Add balance parameters

Implemented the addition of balance parameters to the wallet interface.

🎨 style: Update CSS for the balance page

Made necessary changes to the CSS styles for the balance page of the wallet interface.

🎨 style: Revise style for balance page

Updated the CSS styling for the balance page to enhance its appearance and layout.

✨ feat: Finish balance page

Completed the implementation of the balance page, ensuring all features and functionalities are in place.

🚀 feat: Finalize balance page implementation

Completed the implementation of the balance page and addressed any remaining issues.

🎨 style: Update Figma style

Modified the style of the wallet interface to match the design in Figma.

🎨 style: Update color palette

Adjusted the colors used throughout the wallet interface.

🎉 feat: Finalize wallet design

Completed the design adjustments in style.css, balance.css, and details.css to finalize the wallet interface.